requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
av==16.0.1
backoff==2.2.1
bcrypt==5.0.0
bert-score==0.3.13
bitsandbytes==0.48.2
bm25s==0.2.14
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.3.7
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
contourpy==1.3.3
cuda-bindings==13.0.3
cuda-core==0.3.2
cuda-pathfinder==1.3.2
cuda-python==13.0.3
cuda-toolkit==13.0.0
cudf-cu13==25.10.0
cuml-cu13==25.10.0
cupy-cuda13x==13.6.0
cycler==0.12.1
Cython==3.2.2
dask==2025.9.1
dask-cuda==25.10.0
dask-cudf-cu13==25.10.0
dataclasses-json==0.6.7
datasets==4.4.1
dill==0.4.0
diskcache==5.6.3
distributed==2025.9.1
distributed-ucxx-cu13==0.46.0
distro==1.9.0
durationpy==0.10
einops==0.8.1
evaluate==0.4.6
fastrlock==0.8.3
filelock==3.20.0
flash_attn @ file:///home/rezkykam/SKINspector/3rdparty/flash-attention
flatbuffers==25.9.23
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.12
langsmith==0.4.53
libcudf-cu13==25.10.0
libcuml-cu13==25.10.0
libkvikio-cu13==25.10.0
libraft-cu13==25.10.0
librmm-cu13==25.10.0
libucx-cu13==1.19.0
libucxx-cu13==0.46.0
-e git+https://github.com/linkedin/Liger-Kernel.git@3d075fd7044ae71d5f1e6be2bb25c0cdf838ba18#egg=liger_kernel&subdirectory=../../../3rdparty/Liger-Kernel
llama_cpp_python==0.3.16
llvmlite==0.44.0
locket==1.0.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.6
nltk==3.9.2
numba==0.61.2
numba-cuda==0.19.1
numpy==2.3.5
nvidia-cublas==13.0.0.19
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cccl==13.0.85
nvidia-cuda-cupti==13.0.48
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc==13.0.48
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime==13.0.48
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile==1.15.0.42
nvidia-cufile-cu12==1.13.1.3
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver==12.0.3.29
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse==12.6.2.49
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvtx-cu12==12.8.90
nvidia-nvvm==13.0.88
nvtx==0.2.14
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
partd==1.4.2
peft==0.18.0
pillow==11.3.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pylibcudf-cu13==25.10.0
pylibraft-cu13==25.10.0
pyparsing==3.2.5
pypdf==6.4.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
raft-dask-cu13==25.10.0
rapids-dask-dependency==25.10.0
rapids-logger==0.1.19
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rmm-cu13==25.10.0
rouge_score==0.1.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.2.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
sympy==1.14.0
tblib==3.2.2
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toolz==1.1.0
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
treelite==4.4.1
triton==3.5.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ucxx-cu13==0.46.0
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
zict==3.0.0
zipp==3.23.0
zstandard==0.25.0
